requirements.txt
streamlit==0.66.0
numpy==1.18.5
Keras==2.4.3
Pillow==7.2.0
